requirements.txt
python-dotenv==1.0.0
gitpython==3.1.41
apscheduler==3.10.1
pandas==2.0.3
openpyxl==3.1.2
python-docx==0.8.11
reportlab==4.1.0
jinja2==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5 